Travis help with the Continuous Integration by helping us rebuild and push the image to dockerhub when we commit our code to github. You can set the build to happen when a merge is done to certain branch or whatever you want. it is quite flexible!
Where Travis help with CD is that our Kubernetes we set up will continue to use the latest docker image(s) each time in the pod starts or restarts

So far, We have
Divide an application into microservices. /feed and /user backends are separated into independent projects.
Build and run a container image using Docker. Project includes Dockerfiles to successfully create Docker images for /feed, /user backends, project frontend, and reverse proxy. We use docker-compose build --parallel to build and docker-compose up to run the containers locally.

NOTE:
docker-compose isn't necessarily required for deploying to Kubernetes. However, it is useful because it helps you conceptualize the relationship between the containers and allows you to run all the containers locally.
A container using nginx named reverseproxy is used. It helps add another layer between the frontend and API so that the frontend only uses a single endpoint and doesn't realize it's deployed separately. This is one approach and not necessarily the only way to deploy the services.