Third commit
Add Secret for k8s …
This time we store a file that contains our aws credentials.  We convert the csv file downloaded while creating the IAM user to BASE64
https://base64.guru/converter/encode/file
For this project it is recommended to give it administrator access as the user will need to spin up many resources and for simplicity on the exercise. However, in a real scenario, you should select the appropriate permissions for the user to follow the least privileged security practice recommendation
Setting Volumes makes it easy to see live changes as it happens in the folder path we specify because we create a folder path inside the image with the file. SO in our case we have our aws configured in our docker image just the way it is in our local machine
/.aws is the foder path where our aws files are

Add Secret for K8s …
This secret will store our DB username and password.
The value will be written in Base64 converted from a plaintext
https://base64.guru/converter/encode/text

Add ConfigMap file for k8s
One env-config file which any of the deployment can reference to inside the cluster. In this way, if you make changes to your env variables you only need to change the env-config and each deployment has access to it, without exposing anything
We write plain text values in the file

the livenessProbe in my k8s deployment file …
The time i set initially was too short. My container has not started my code, and the endpoint for health is not yet ready. So this causes my app to enter CrashLoopBackOff
You have more ways to use livenessProbe or even readinessProbe
Check out:   https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/